Move "hydrate" modules to "boot"
